Add a gogocompat package for GoGo compatibility shims.
Helpful for Envoy config questions (e.g. issue #40).
